Moved iter to file.h
The old impl is now using FileIter

+++ Lib/arena.c
@@ -48,8 +48,8 @@ void* arena_realloc(Arena* arena, u8* ptr, isize current_size, ptrdiff_t new_siz
   isize align = _Alignof(u8);
   isize padding = -(isize)(arena->data + arena->offset) & (align - 1);
 
-  isize old_end = ptr + current_size;
-  isize arena_end = arena->data + arena->offset;
+  isize old_end = (isize)ptr + current_size;
+  isize arena_end = (isize)arena->data + arena->offset;
 
 
   isize available = arena->cap - arena->offset - padding - new_size;
